pyproject.toml
[tool.poetry]
name = "techstore"
version = "0.1.0"
description = "TechStore SaaS - FastAPI + HTMX"
authors = ["Lucas <turlettilucasdev@gmail.com>"]
readme = "README.md"
packages = [{include = "app", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
jinja2 = "^3.1.2"
python-multipart = "^0.0.6"
sqlalchemy = "^2.0.23"
alembic = "^1.12.1"
psycopg2-binary = "^2.9.9"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-dotenv = "^1.0.0"
pydantic-settings = "^2.1.0"
aiofiles = "^23.2.0"
httpx = "^0.25.2"
python-slugify = "^8.0.1"
email-validator = "^2.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
black = "^23.11.0"
ruff = "^0.1.6"
pre-commit = "^3.5.0"
factory-boy = "^3.3.0"
coverage = "^7.3.2"
pytest-cov = "^4.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py311"
line-length = 88
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long
    "B008",  # do not perform function calls in argument defaults
    "E402",  # module level import not at top of file (for scripts)
    "UP007",  # Use X | Y for type annotations (requires Python 3.10+)
]
per-file-ignores = {"scripts/*" = ["E402", "I001"]}
exclude = [
    ".bmad-core",
    ".claude",
    ".cursor",
    "docs",
    ".git",
    ".pre-commit",
    ".ruff_cache",
    ".pytest_cache",
    "alembic/versions",
    "migrations",
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = ["tests"]
